Fix 1 km SensorZenith dtype in tests

+++ satpy/tests/reader_tests/test_mersi_l1b.py
@@ -207,7 +207,7 @@ def _get_1km_data(num_scans, rows_per_scan, num_cols, filetype_info):
                     dims=("_ir250_bands", "_rows", "_cols")),
         f"{key_prefix}SensorZenith":
                 xr.DataArray(
-                    da.ones((num_scans * rows_per_scan, num_cols), chunks=1024),
+                    da.ones((num_scans * rows_per_scan, num_cols), chunks=1024, dtype=np.int16),
                     attrs={
                         "Slope": np.array([.01] * 1, dtype=np.float32),
                         "Intercept": np.array([0.] * 1, dtype=np.float32),
